refactor(sync): remove legacy sync-v1.0
