+++ tests/others/test_init_manager.py
@@ -50,7 +50,6 @@ class SimpleManagerInitializationTestCase(unittest.TestCase):
             builder = TestBuilder()
             builder.set_tx_storage(self.tx_storage)
             builder.disable_sync_v1()
-            builder.disable_sync_v1_1()
             builder.disable_sync_v2()
             builder.build()
